requirements.txt
py-cord==2.5.0
aiohttp==3.9.5
python-dotenv==1.0.1
